% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{expect_lol}
\alias{expect_lol}
\title{Test for lol}
\usage{
expect_lol(current)
}
\arguments{
\item{current}{the output to be tested.}
}
\description{
Test if a value is equal to \code{"lol"}
}
\examples{
library(tinytest)
expect_lol("lol")
expect_lol("lulz")

}
